backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.1
pydantic-settings==2.8.1
supabase==2.15.0
psycopg[binary]==3.2.6
sqlalchemy==2.0.40
alembic==1.15.2
pandas==2.2.3
polars==1.26.0
pyarrow==19.0.1
openpyxl==3.1.5
httpx==0.28.1
tenacity==9.0.0
python-dotenv==1.1.0
PyJWT==2.9.0
sse-starlette==2.2.1
workalendar==17.0.0
openai==1.69.0
